Разделил на классы в отдельных файлах. Добавил соль и хэш в аутентификации
